Move analytics code to tracking (ex-tracker) package
