+++ kifi-backend/modules/search/app/com/keepit/search/controllers/search/SearchCacheController.scala
@@ -4,7 +4,7 @@ import com.google.inject.Inject
 import play.api.mvc.Action
 import com.keepit.common.controller.SearchServiceController
 import com.keepit.search._
-import com.keepit.search.tracker.S3BackedResultClickTrackerBuffer
+import com.keepit.search.tracking.S3BackedResultClickTrackerBuffer
 
 class SearchCacheController @Inject() (
   s3BackedResultClickTrackerBuffer: S3BackedResultClickTrackerBuffer)
