+++ kifi-backend/modules/search/app/com/keepit/search/controllers/search/SearchEventController.scala
@@ -3,12 +3,9 @@ package com.keepit.search.controllers.search
 import com.google.inject.Inject
 import com.keepit.common.controller.SearchServiceController
 import com.keepit.common.db.Id
-import com.keepit.model.{ User, NormalizedURI }
-import com.keepit.search._
-import play.api.libs.json._
+import com.keepit.model.User
+import com.keepit.search.tracking.SearchEventCommander
 import play.api.mvc.Action
-import com.keepit.common.akka.SafeFuture
-import play.api.libs.concurrent.Execution.Implicits._
 
 class SearchEventController @Inject() (searchEventCommander: SearchEventCommander) extends SearchServiceController {
