+++ kifi-backend/modules/search/app/com/keepit/search/tracking/ClickHistoryTracker.scala
@@ -1,4 +1,4 @@
-package com.keepit.search.tracker
+package com.keepit.search.tracking
 
 import com.keepit.common.akka.SafeFuture
 import com.keepit.common.db.Id
