+++ kifi-backend/modules/search/app/com/keepit/search/tracking/ProbablisticLRU.scala
@@ -1,4 +1,4 @@
-package com.keepit.search.tracker
+package com.keepit.search.tracking
 
 import com.keepit.common.logging.Logging
 import com.keepit.common.cache.TransactionalCaching.Implicits.directCacheAccess
