+++ kifi-backend/modules/search/app/com/keepit/search/tracking/ProbablisticLRUChunkCache.scala
@@ -1,4 +1,4 @@
-package com.keepit.search.tracker
+package com.keepit.search.tracking
 
 import com.keepit.common.logging.AccessLog
 import com.keepit.search.{ FullFilterChunkId }
